feat: 增加sys_exec多系统支持
