+++ sys_exec/sys_exec_darwin.go
@@ -1,0 +1,19 @@
+//go:build darwin
+// +build darwin
+
+package sys_exec
+
+func BuildCmd(cmdStr string) *exec.Cmd {
+	return exec.Command("sh", "-c", cmdStr)
+}
+
+func SetPgid(cmd *exec.Cmd) {
+	if cmd == nil {
+		return
+	}
+	if cmd.SysProcAttr == nil {
+		cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
+	} else {
+		cmd.SysProcAttr.Setpgid = true
+	}
+}
